requirements.txt
Flask==3.1.2
openai==1.99.5
python-dotenv==1.1.1